Implement .mbedignore loader
This class extends MbedIgnoreSet from MBED OS Python API. I extended it
because it already has `.mbedignore` loader functions, but the class
doesn't expose any method to get the collected patterns.

The class has methods that are functions with side effects. This is not
clean, but there is no clean way to utilize the MbedIgnoreSet methods
without side effects. The used MbedIgnoreSet.add_mbedignore() function
mutates self._ignore_patterns, so performing any other operations on
that member will, of course, have side effects.

The collect_ignore_patterns() implement caching. When the ignore
patterns has been collected once, they won't be recollected. The result
from the first collection is stored internally and returned on each
consecutive call to the method.

The result of collect_ignore_patterns() is supposed to be provided to
Resources.add_ignore_patterns() function. To be more specific, this
method will be called on the extender of Resources, which is
MbedResourcesFixedPath. The function expects patterns in specific
format, I explain it along with
_remove_first_directory_name_from_patterns() below.

The methods following methods mutate self._ignore_patterns:

* _remove_mbed_framework_prefix_from_patterns() - performs filtering
which removes 'mbed-os/' or 'framework-mbed/' prefix from the ignore
patterns.
* _remove_first_directory_name_from_patterns() - removes first directory
name from the patterns. This comes from experimenting. I have checked
what patterns does Resources.add_ignore_patterns() expect. The patterns
are only obeyed (directories are ignored) when there is no first folder
(directory) name in the pattern. So, e.g. 'features/cellular/*' will not
be properly applied, but 'cellular/*' will. Finally, the method from
this bullet, will turn the pattern 'features/cellular/*' to
'cellular/*'.