Load .mbedignore patterns from root project dir
This will load `.mbedignore` patterns from the root directory of the
project. It will use the PioMbedignoreLoader to accomplish that. The
ignore patterns will be provided to the MbedResourcesFixedPath before
the sources are collected.

The first attempt to accomplish this was through the MBED OS API,
specifically 'tools.build_api.prepare_toolchain', using the 'ignore'
parameter, according to the answer:
https://community.platformio.org/t/eager-to-implement-handling-of-mbedignore-within-pio/23735/2
to my post. Unfortunately it is not possible, because no
'add_ignore_patterns' is implemented in:

TOOLCHAIN_CLASSES = {
    u'ARM': arm.ARM_STD,
    u'uARM': arm.ARM_MICRO,
    u'ARMC6': arm.ARMC6,
    u'GCC_ARM': gcc.GCC_ARM,
    u'IAR': iar.IAR
}

This is a bug in mbed, I have raised an issue:
https://github.com/ARMmbed/mbed-os/issues/15125