Remove match all patterns
There was a defect, when one wanted to e.g. exclude:

    UNITTESTS/*

Then after removing the first folder name, there is only:

    *

left - so "match all". We have to filter out those records and document
that the users can't use paths with one directory. It means that it is
not OK, and will be skipped:

    UNITTESTS/*

Users shall use instead:

    UNITTESTS/stubs/*
    UNITTESTS/features/*
    ...

(BTW) UNITTESTS won't be compiled on 'pio run' anyway ;)